+++ packages/studio-server/package.json
@@ -25,6 +25,7 @@
 		"remotion": "workspace:*",
 		"@remotion/bundler": "workspace:*",
 		"@remotion/renderer": "workspace:*",
+		"@remotion/studio-shared": "workspace:*",
 		"memfs": "3.4.3",
 		"source-map": "0.7.3",
 		"open": "^8.4.2"
