--- packages/studio-server/src/required-chromium-options.ts
@@ -1,5 +1,0 @@
-import type {ChromiumOptions, OpenGlRenderer} from '@remotion/renderer';
-
-export type RequiredChromiumOptions = Required<ChromiumOptions>;
-
-export type UiOpenGlOptions = OpenGlRenderer | 'default';
