+++ packages/studio-shared/.eslintrc
@@ -1,0 +1,13 @@
+{
+	"extends": "@jonny",
+	"rules": {
+		"no-console": "error",
+		"no-restricted-imports": [
+			"error",
+			{
+				"patterns": ["@remotion/*/src/*", "remotion/src/*"]
+			}
+		],
+		"@typescript-eslint/no-use-before-define": "off"
+	}
+}
