+++ packages/studio-shared/.npmignore
@@ -1,0 +1,5 @@
+src
+tsconfig.json
+dist/test
+.eslintrc.js
+.prettierrc.js
