+++ packages/studio-shared/package.json
@@ -1,7 +1,7 @@
 {
-	"name": "@remotion/studio",
+	"name": "@remotion/studio-shared",
 	"version": "4.0.88",
-	"description": "Remotion Editor",
+	"description": "Shared items between Remotion Studio frontend and backend",
 	"main": "dist",
 	"sideEffects": false,
 	"scripts": {
@@ -20,45 +20,23 @@
 	"bugs": {
 		"url": "https://github.com/remotion-dev/remotion/issues"
 	},
-	"dependencies": {
-		"semver": "7.5.3",
-		"remotion": "workspace:*",
-		"@remotion/bundler": "workspace:*",
-		"@remotion/player": "workspace:*",
-		"@remotion/media-utils": "workspace:*",
-		"@remotion/renderer": "workspace:*",
-		"memfs": "3.4.3",
-		"source-map": "0.7.3",
-		"open": "^8.4.2"
-	},
+	"dependencies": {},
 	"devDependencies": {
-		"react": "18.2.0",
-		"react-dom": "18.2.0",
 		"@jonny/eslint-config": "3.0.276",
 		"@types/node": "18.14.6",
-		"@types/react": "18.0.26",
-		"@types/react-dom": "18.0.11",
 		"eslint": "8.42.0",
 		"eslint-plugin-10x": "1.5.2",
 		"eslint-plugin-react": "7.32.2",
 		"eslint-plugin-react-hooks": "4.4.0",
 		"prettier": "3.1.1",
-		"@types/semver": "^7.3.4",
 		"prettier-plugin-organize-imports": "3.2.4",
-		"vitest": "0.31.1",
-		"zod": "^3.22.3",
-		"@remotion/zod-types": "workspace:*"
+		"@remotion/renderer": "workspace:*"
 	},
 	"publishConfig": {
 		"access": "public"
 	},
 	"exports": {
-		".": {
-			"import": "./dist/index.js",
-			"require": "./dist/index.js",
-			"types": "./dist/index.d.js"
-		},
-		"./package.json": "./package.json",
-		"./entry": "./dist/previewEntry.js"
+		".": "./dist/index.js",
+		"./package.json": "./package.json"
 	}
 }
